Add custom name to columns of product abd category table

+++ InventoryManagmentSystem/src/main/java/com/kanak/ims/controller/ProductController.java
@@ -73,7 +73,7 @@ public class ProductController {
     }
 
     @GetMapping("/byCategory/{cat}")
-    public List<Product> findByCategoryType(@PathVariable int cat){
+    public List<Product> findByCategoryType(@PathVariable Long cat){
         return productService.findByCategoryType(cat);
     }
